Track channel download rate in bytes per second.

+++ include/bitcoin/node/protocols/protocol.hpp
@@ -42,7 +42,7 @@ protected:
     }
 
     /// Report performance, false directs self-terminate.
-    void performance(uint64_t channel, size_t speed,
+    void performance(uint64_t channel, uint64_t speed,
         network::result_handler&& handler) const NOEXCEPT;
 
     /// Configuration settings for all libraries.
